possibly fixes #15
The error will have to do with PR #14. As I am still very limited in time, I gave it a shot with this fix, please try if it works now for you.
Otherwise maybe @wmattei could have a look at it? Also a repro would help, as I cannot reproduce it here.
Cheers!